organized components into atomic design folders
