+++ frontend-new/src/components/atoms/SectionDivider.tsx
