+++ frontend-new/src/components/atoms/buttons/JoinWaitlistButton.tsx
