+++ frontend-new/src/screens/TasksScreen.tsx
@@ -5,10 +5,10 @@ import Cookies from 'js-cookie'
 import React, { useEffect, useRef, useState } from 'react'
 import BottomSheet from 'reanimated-bottom-sheet'
 import { findTaskById } from '../utils/task'
-import EditSheet from '../components/tasks/EditSheet'
-import CreateNewTask from '../components/tasks/CreateNewTask'
-import TasksScreenHeader from '../components/tasks/Header'
-import TaskSections from '../components/tasks/Sections'
+import EditSheet from '../components/organisms/EditSheet'
+import CreateNewTask from '../components/organisms/CreateNewTask'
+import TasksScreenHeader from '../components/molecules/Header'
+import TaskSections from '../components/organisms/Sections'
 import { useGetTasksQuery } from '../services/generalTaskApi'
 import { Screens, Flex, Colors, Dimensions } from '../styles'
 import { useAppDispatch } from '../redux/hooks'
